Added the solution for "Extra utmaning" Added a abstract class Shape as the Super Class Made the getArea, getCircumference and getVolume abstract methods requiring the derived classes to implement these methods with the ´override´ keyword and define it for the specific class
Changed the private attributes to protected and only included the _Height as a double in the Shape class the other subclasses dont require the same attributes besides _Height so it felt like a waste to implement attributes in the SuperClass that would not be used.
The access modifier protected ensures that the attributes stays encapsulated inside the classes and cannot get modified outside of the classes. Where it was needed I added public geters.

public class Circle : Shape makes the Circle class inherit the Shape class
public class Triangle : Shape makes the Triangle class inherit the Shape class